:pencil2: rename: itemdao -> repository로 변경
