+++ src/main/java/com/luckyGirls/forYourNutrition/dao/repository/IRecommendRepository.java
@@ -1,4 +1,4 @@
-package com.luckyGirls.forYourNutrition.dao;
+package com.luckyGirls.forYourNutrition.dao.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.luckyGirls.forYourNutrition.domain.IRecommend;
 
 @Repository
-public interface IRecommendDao extends JpaRepository<IRecommend, Long> {
+public interface IRecommendRepository extends JpaRepository<IRecommend, Long> {
 }
