+++ src/main/java/com/luckyGirls/forYourNutrition/dao/repository/ItemRepository.java
@@ -1,10 +1,11 @@
-package com.luckyGirls.forYourNutrition.dao;
+package com.luckyGirls.forYourNutrition.dao.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.luckyGirls.forYourNutrition.domain.Item;
 
-public interface ItemDao {
+@Repository
+public interface ItemRepository extends JpaRepository<Item, Long> {
 
 }
