+++ src/main/java/com/luckyGirls/forYourNutrition/service/IRecommendService.java
@@ -3,8 +3,8 @@ package com.luckyGirls.forYourNutrition.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.luckyGirls.forYourNutrition.dao.IRecommendDao;
-import com.luckyGirls.forYourNutrition.dao.ItemDao;
+import com.luckyGirls.forYourNutrition.dao.repository.IRecommendRepository;
+import com.luckyGirls.forYourNutrition.dao.repository.ItemRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -12,6 +12,6 @@ import lombok.RequiredArgsConstructor;
 @Transactional(readOnly = true)
 @Service
 public class IRecommendService {
-	private static ItemDao itemDao;
-	private static IRecommendDao iRecommendDao;
+	private static ItemRepository itemRepository;
+	private static IRecommendRepository iRecommendRepository;
 }
